+++ intellij_project/src/main/java/it/polimi/telco_webapp/services/OptionsOrderedPricesViewService.java
@@ -7,15 +7,17 @@ import jakarta.ejb.Stateless;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 
+import java.math.BigDecimal;
+
 @Stateless(name = "OptionsOrderedPricesView")
 public class OptionsOrderedPricesViewService {
     @PersistenceContext(unitName = "telco_webapp")
     private EntityManager em;
     public OptionsOrderedPricesViewService() {}
 
-    public float getOptionsPrice(int order_id) {
+    public BigDecimal getOptionsPrice(int order_id) {
         OptionsOrderedPricesView optionsPrices = em.find(OptionsOrderedPricesView.class, order_id);
-        float price = optionsPrices.getSumOfSales();
+        BigDecimal price = optionsPrices.getSumOfSales();
         return price;
     }
 }
